Got mongodb connecting and further refined app formatting
